# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Absensi API
  version: 0.2.0
  description: >
    API untuk aplikasi absensi dengan QR (NestJS + MariaDB).
    Pembaruan: multi-lokasi dari awal, shift tetap per karyawan, payroll harian.
servers:
  - url: http://localhost:3000
    description: Local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiError:
      type: object
      properties:
        statusCode: { type: integer }
        message: { type: string }
        error: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
    LoginResponse:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        user:
          $ref: '#/components/schemas/User'
...
    User:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        employee:
          $ref: '#/components/schemas/Employee'
        roles:
          type: array
          items: { type: string }
    Employee:
      type: object
      properties:
        id: { type: integer }
        full_name: { type: string }
        role_title: { type: string, nullable: true }
        gender: { type: string, enum: [M, F], nullable: true }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        join_date: { type: string, format: date, nullable: true }
        is_active: { type: boolean }
        default_location_id: { type: integer, nullable: true }
        default_shift_id: { type: integer, nullable: true }
        base_salary_rate: { type: number, nullable: true }
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
...
  /auth/change-password:
    post:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      summary: Ganti password akun saat ini
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string }
                new_password: { type: string, minLength: 6 }
      responses:
        '200': { description: OK }
  /me:
    get:
      tags: [Me]
      security: [{ bearerAuth: [] }]
      summary: Profile saat ini
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
  /me/identity:
    patch:
      tags: [Me]
      security: [{ bearerAuth: [] }]
      summary: Update nama lengkap dan username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name: { type: string }
                username: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
  /me/phone:
    patch:
      tags: [Me]
      security: [{ bearerAuth: [] }]
      summary: Update nomor HP (employee)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
...